# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Literary Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{username}:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get user profile by username
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/me/profile:
    patch:
      operationId: updateMyProfile
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/me/avatar:
    post:
      operationId: uploadAvatar
      tags: [users]
      summary: Upload avatar (base64)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AvatarUpload"
      responses:
        "200":
          description: Avatar updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /links:
    get:
      operationId: getMyLinks
      tags: [links]
      summary: Get current user links
      responses:
        "200":
          description: Links list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Link"
    post:
      operationId: createLink
      tags: [links]
      summary: Create a link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkInput"
      responses:
        "201":
          description: Link created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Link"
  /links/{id}:
    patch:
      operationId: updateLink
      tags: [links]
      summary: Update a link
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkUpdate"
      responses:
        "200":
          description: Link updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Link"
    delete:
      operationId: deleteLink
      tags: [links]
      summary: Delete a link
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Link deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /users/{username}/books:
    get:
      operationId: getAuthorBooks
      tags: [books]
      summary: Get books by a specific author
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Author books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Book"
  /users/{username}/links:
    get:
      operationId: getAuthorLinks
      tags: [links]
      summary: Get links for a specific author profile
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Author links
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Link"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
...
    User:
      type: object
      required: [id, username, displayName, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    UserProfile:
      type: object
      required: [id, username, displayName, createdAt, books, links]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        books:
          type: array
          items:
            $ref: "#/components/schemas/Book"
        links:
          type: array
          items:
            $ref: "#/components/schemas/Link"
    ProfileUpdate:
      type: object
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 60
        bio:
          type: string
          maxLength: 300
    AvatarUpload:
      type: object
      required: [dataUrl]
      properties:
        dataUrl:
          type: string
    Link:
      type: object
      required: [id, title, url, sortOrder]
      properties:
        id:
          type: integer
        title:
          type: string
        url:
          type: string
        icon:
          type: ["string", "null"]
        sortOrder:
          type: integer
    LinkInput:
      type: object
      required: [title, url]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 80
        url:
          type: string
        icon:
          type: string
    LinkUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 80
        url:
          type: string
        icon:
          type: string
        sortOrder:
          type: integer
    Book:
      type: object
      required: [id, title, authorId, authorUsername, authorDisplayName, genre, status, isAdult, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        genre:
          type: string
        status:
          type: string
          enum: [draft, published]
        isAdult:
          type: boolean
        authorId:
          type: integer
        authorUsername:
          type: string
        authorDisplayName:
          type: string
        chapterCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...