# lib/api-zod/src/generated/types/link.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Literary Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Link {
  id: number;
  title: string;
  url: string;
  /** @nullable */
  icon?: string | null;
  sortOrder: number;
}